requirements.txt
# AI Code Review Bot Dependencies
# Python 3.11+

# HTTP client for Bitbucket API calls
requests==2.31.0

# OpenAI official Python client
openai==1.54.4

# HTTPX version compatible with OpenAI client
httpx==0.27.2

# YAML configuration parser
PyYAML==6.0.1

# Environment variable management (optional but recommended)
python-dotenv==1.0.0
